requirements.txt
annotated-types==0.7.0
anyio==4.4.0
av==12.1.0
certifi==2024.6.2
charset-normalizer==3.3.2
coloredlogs==15.0.1
ctranslate2==4.3.1
distro==1.9.0
faster-whisper==1.0.2
ffmpeg-python==0.2.0
filelock==3.14.0
flatbuffers==24.3.25
fsspec==2024.6.0
future==1.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.3
humanfriendly==10.0
idna==3.7
mpmath==1.3.0
numpy==1.26.4
onnxruntime==1.18.0
openai==1.33.0
packaging==24.1
protobuf==5.27.1
pydantic==2.7.3
pydantic_core==2.18.4
PyYAML==6.0.1
requests==2.32.3
setuptools==70.0.0
sniffio==1.3.1
sympy==1.12.1
tokenizers==0.19.1
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.1
